+++ GEECS-Data-Utils/geecs_data_utils/data/__init__.py
@@ -12,6 +12,12 @@ from geecs_data_utils.data.cleaning import (
     sigma_clip_frame,
     sigma_nan_frame,
 )
+from geecs_data_utils.data.binning import (
+    BinnedFrame,
+    BinningConfig,
+    bin_frame,
+    compute_bin_key,
+)
 from geecs_data_utils.data.row_filters import (
     FilterCondition,
     FilterGroup,
@@ -31,6 +37,10 @@ from geecs_data_utils.data.columns import (
 from geecs_data_utils.data.dataset import DatasetBuilder, DatasetFrame, LoadScansReport
 
 __all__ = [
+    "BinnedFrame",
+    "BinningConfig",
+    "bin_frame",
+    "compute_bin_key",
     "FilterCondition",
     "FilterGroup",
     "RowFilters",